% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes.r
\name{converter}
\alias{converter}
\title{converter}
\usage{
converter(valor, origem, destino)
}
\arguments{
\item{valor}{Codigo do municipio}

\item{origem}{Base do codigo original}

\item{destino}{Base do codigo desejado}
}
\value{
Retorna o valor convertido para a base destino
}
\description{
Converte o codigo da base original para uma base desejada
}
