% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes.r
\docType{data}
\name{empilhados}
\alias{empilhados}
\title{Dados empilhados}
\format{
Uma lista que descreve as colunas:
\describe{
\item{uf}{Unidade federativa - Estado do municipio}
\item{municipio}{Nome do municipio sem espacos e em maiusculo}
\item{cod_ibge_resumido}{Codigo do ibge resumido}
\item{cod_ibge_completo}{Codigo do ibge completo}
\item{nome_original}{Nome do municipio sem modificacao}
\item{cod_sef_mg}{Codigo SEF-MG}
\item{cod_tom_serpro}{Codigo Tom Serpro}
\item{cod_tse}{Codigo TSE}
}
}
\source{
Dados retirados de fontes oficiais e compilados em uma unica base.
}
\usage{
empilhados
}
\description{
Base de dados referentes a municipios brasileiros e seus respectivos codigos de identificacao em diferentes bases
}
\keyword{datasets}
