% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes.r
\name{municipio}
\alias{municipio}
\title{municipio}
\usage{
municipio(valor, base)
}
\arguments{
\item{valor}{Codigo do municipio}

\item{base}{Base de origem}
}
\value{
Retorna o nome do municipio
}
\description{
Retorna o nome do municipio a partir do codigo e base original
}
